feat: Add video alignment and refined processing scripts
This commit introduces a new video alignment script () that estimates and corrects train video tilt. It also adds  to move processed files to a date named folder in . Further, the processing script () is modified to work on already aligned videos. This includes adding  to be ran before, and removing the alignment functionality, to increase speed and reliability. This improves the overall data processing pipeline.